feat(core): headless execution; argument handling
Currently this is only tested on Linux. Will test OS X next and
currently no plan of testing this on Window...

How to use the feature:

1. When building with CMake, you can use `cmake -DBUILD_HEADLESS=on ..`
which is equivalent of `#define HEADLESS` and the main file will be
compile as dictated.

2. When using Xcode or other IDE, users can directly edit `main.cpp`
file by adding the defined variable `HEADLESS`.

To run the program, there are different usage on different platform.

- OS X   : open -n ./ESP.app --args <ESP session directory>
- Linux  : ./ESP <ESP session directory>
- Window : ???

Note the session directory should be absolute path.

References #369 Running on CHIP/PI
References #298 Separate ESP from ofApp (GUI)
Closes #205 Command line prediction mode.